doing class stuff
I made css into scss by adding an _ before the name and a .scss at the end. It automatically converts it to scss. sass is old sass with no {. I relinked my images and links for a stylesheet. I change the image folder name to imgs and the components folder name to public. Then I copied the scss files back into the public files with the html so they could actually display my style even though that is not what Zack said to do.